force rename of adapter
