+++ src/adapter/HtmlWebpackPluginn.ts
